monitoring/requirements.txt
aiogram==3.7.0
telethon==1.34.0
aiosqlite==0.20.0
python-dotenv==1.0.1
cryptg==0.4.0
Pillow==10.3.0
aiofiles==23.2.1
